requirements.txt
requests==2.31.0
apscheduler==3.10.1
python-telegram-bot==20.5
twilio==8.3.0
python-dotenv==1.0.0
PyYAML==6.0.1
validators==0.22.0
python-dateutil==2.8.2
sqlalchemy==2.0.23
flask==3.0.0
gunicorn==21.2.0
prometheus-client==0.19.0